refactor(room-type): Improves code organization
